refactor some events out into own classes
