+++ src/bunit.web/RenderingPort/Events/EventDelegator.cs
@@ -325,191 +325,3 @@ internal class EventDelegator
 		}
 	}
 }
-
-
-internal sealed class EventHandlerInfosForElement
-{
-	// Although we *could* track multiple event handlers per (element, eventName) pair
-	// (since they have distinct eventHandlerId values), there's no point doing so because
-	// our programming model is that you declare event handlers as attributes. An element
-	// can only have one attribute with a given name, hence only one event handler with
-	// that name at any one time.
-	// So to keep things simple, only track one EventHandlerInfo per (element, eventName)
-	private readonly Dictionary<string, EventHandlerInfo> handlers = new();
-	private readonly Dictionary<string, bool> preventDefaultFlags = new();
-	private readonly Dictionary<string, bool> stopPropagationFlags = new();
-
-	public EventHandlerInfosForElement()
-	{
-	}
-
-	internal EventHandlerInfo? GetHandler(string eventName)
-	{
-		return handlers.TryGetValue(eventName, out var result)
-			? result
-			: default;
-	}
-
-	internal void SetHandler(string eventName, EventHandlerInfo handler)
-	{
-		handlers.Add(eventName, handler);
-	}
-
-	internal void RemoveHandler(string eventName)
-	{
-		handlers.Remove(eventName);
-	}
-
-	internal bool PreventDefault(string eventName, bool? setValue = null)
-	{
-		if (setValue is not null)
-		{
-			preventDefaultFlags[eventName] = setValue.Value;
-		}
-
-		return preventDefaultFlags.ContainsKey(eventName)
-			? preventDefaultFlags[eventName]
-			: false;
-	}
-
-	internal bool StopPropagation(string eventName, bool? setValue = null)
-	{
-		if (setValue is not null)
-		{
-			stopPropagationFlags[eventName] = setValue.Value;
-		}
-
-		return stopPropagationFlags.ContainsKey(eventName)
-			? stopPropagationFlags[eventName]
-			: false;
-	}
-}
-
-internal sealed class EventHandlerInfo
-{
-	public IElement Element { get; }
-	public string EventName { get; }
-	public ulong EventHandlerId { get; set; }
-	// The component whose tree includes the event handler attribute frame, *not* necessarily the
-	// same component that will be re-rendered after the event is handled (since we re-render the
-	// component that supplied the delegate, not the one that rendered the event handler frame)
-	public int RenderingComponentId { get; }
-
-	public EventHandlerInfo(IElement element, string eventName, ulong eventHandlerId, int renderingComponentId)
-	{
-		Element = element;
-		EventName = eventName;
-		EventHandlerId = eventHandlerId;
-		RenderingComponentId = renderingComponentId;
-	}
-}
-
-internal class EventInfoStore
-{
-	private readonly Dictionary<ulong, EventHandlerInfo> infosByEventHandlerId = new();
-	private readonly Dictionary<string, int> countByEventName = new();
-	private readonly DomEventHandler globalListener;
-	private readonly IDocument document;
-
-	public EventInfoStore(DomEventHandler globalListener, IDocument document)
-	{
-		this.globalListener = globalListener;
-		this.document = document;
-		EventNameAliasRegisteredCallbacks.Add(HandleEventNameAliasAdded);
-	}
-
-	internal void Add(EventHandlerInfo info)
-	{
-		if (infosByEventHandlerId.ContainsKey(info.EventHandlerId))
-		{
-			// Should never happen, but we want to know if it does
-			throw new ArgumentException($"Event {info.EventHandlerId} is already tracked");
-		}
-
-		infosByEventHandlerId.Add(info.EventHandlerId, info);
-
-		AddGlobalListener(info.EventName);
-	}
-
-	internal EventHandlerInfo Get(ulong eventHandlerId)
-		=> infosByEventHandlerId[eventHandlerId];
-
-	internal void AddGlobalListener(string eventName)
-	{
-		// If this event name is an alias, update the global listener for the corresponding browser event
-		eventName = GetBrowserEventName(eventName);
-
-		if (countByEventName.ContainsKey(eventName))
-		{
-			countByEventName[eventName] += 1;
-		}
-		else
-		{
-			countByEventName[eventName] = 1;
-
-			// To make delegation work with non-bubbling events, register a 'capture' listener.
-			// We preserve the non-bubbling behavior by only dispatching such events to the targeted element.
-			var useCapture = NonBubblingEvents.Contains(eventName);
-			document.AddEventListener(eventName, globalListener, useCapture);
-		}
-	}
-
-	internal void Update(ulong oldEventHandlerId, ulong newEventHandlerId)
-	{
-		if (infosByEventHandlerId.ContainsKey(newEventHandlerId))
-		{
-			// Should never happen, but we want to know if it does
-			throw new ArgumentException($"Event {newEventHandlerId} is already tracked");
-		}
-
-		// Since we're just updating the event handler ID, there's no need to update the global counts
-		var info = infosByEventHandlerId[oldEventHandlerId];
-		infosByEventHandlerId.Remove(oldEventHandlerId);
-		info.EventHandlerId = newEventHandlerId;
-		infosByEventHandlerId[newEventHandlerId] = info;
-	}
-
-	internal EventHandlerInfo? Remove(ulong eventHandlerId)
-	{
-		if (infosByEventHandlerId.TryGetValue(eventHandlerId, out var info))
-		{
-			infosByEventHandlerId.Remove(eventHandlerId);
-
-			// If this event name is an alias, update the global listener for the corresponding browser event
-			var eventName = GetBrowserEventName(info.EventName);
-
-			if (countByEventName.TryGetValue(eventName, out var count))
-			{
-				if (--count == 0)
-				{
-					countByEventName.Remove(eventName);
-					document.RemoveEventListener(eventName, globalListener);
-				}
-				else
-				{
-					countByEventName[eventName] = count;
-				}
-			}
-		}
-
-		return info;
-	}
-
-	private void HandleEventNameAliasAdded(string aliasEventName, string browserEventName)
-	{
-		// If an event name alias gets registered later, we need to update the global listener
-		// registrations to match. This makes it equivalent to the alias having been registered
-		// before the elements with event handlers got rendered.
-		if (countByEventName.TryGetValue(aliasEventName, out var countByAliasEventName))
-		{
-			// Delete old
-			countByEventName.Remove(aliasEventName);
-			document.RemoveEventListener(aliasEventName, globalListener);
-
-			// Ensure corresponding count is added to new
-			AddGlobalListener(browserEventName);
-			countByEventName[browserEventName] += countByAliasEventName - 1;
-		}
-	}
-
-}
